Merge in EDGEAI-ALGO/edgeai-mmdetection3d from develop_vad_far3d_onnx to develop

* commit 'b84eedcc345993f285d0b82c131ae67cdfb06e9a':
  [EDGEAI_ALGO-156] 1. Sparse4D onnx model export. 2. Update setup.sh to compile Sparse4D custom ops
  [EDGEAI_ALGO-156] Support Sparse4D inference. For tracking metrics, - Updated nuscenes-devkit==1.2.0 along with - motmetrics==1.4.0
  Moved GlobalRotScaleTransImage and ResizeCropFlipImage, which are used by multiple models to projects_edgeai/edgeai_mmdet3d.
  [EDGEAI_ALGO-156] Added Sparse4D. 1. Updated pickle data file generation. 2. Sparse4D module is added into project_edgeai.
  changes made in pkl file generation for sparse 4d
  number of prev key frame changed to 1
  Update update_infos_to_v2.py: Check existences of frame_idx, prev, etc.
  [VAD] EDGEAI_ALGO-170: Updated VAD onnx model so that it runs correctly in onnxrt.

+++ projects_edgeai/edgeai_mmdet3d/transformers/bevformer_encoder.py
@@ -68,7 +68,7 @@ class BEVFormerEncoder(TransformerLayerSequence):
             ref_3d = torch.stack((xs, ys, zs), -1)
             ref_3d = ref_3d.permute(0, 3, 1, 2).flatten(2).permute(0, 2, 1)
             ref_3d = ref_3d[None].repeat(bs, 1, 1, 1)
-            return ref_3d
+            return ref_3d.to(torch.float32)
 
         # reference points on 2D bev plane, used in temporal self-attention (TSA).
         elif dim == '2d':
@@ -82,7 +82,9 @@ class BEVFormerEncoder(TransformerLayerSequence):
             ref_x = ref_x.reshape(-1)[None] / W
             ref_2d = torch.stack((ref_x, ref_y), -1)
             ref_2d = ref_2d.repeat(bs, 1, 1).unsqueeze(2)
-            return ref_2d
+            # Sometimes, ref_2d is exported to float64 even though it is float32
+            # Therefore, type cast to float32 explicitly
+            return ref_2d.to(torch.float32)
 
     # This function must use fp32!!!    
     def point_sampling(self, reference_points, pc_range,  img_metas):
